+++ test/components/barrel.test.tsx
@@ -40,6 +40,7 @@ describe('components barrel', () => {
       'useLocalStorage',
       'useMovable',
       'useProcessingText',
+      'usePropState',
       'useRefFunction',
       'useRefValue',
       'useRowSelection',